infra/requirements.txt
aws-cdk-lib>=2.145.0
constructs>=10.3.0



